requirements.txt
openai==1.63.2
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
streamlit==1.42.1
